refactor b2b tabs management
