+++ src/PrestaShopBundle/Resources/config/services/bundle/event_listener.yml
@@ -1,6 +1,7 @@
 services:
   _defaults:
     public: true
+    autowire: true
 
   prestashop.dispatcher.legacy_hooks.subscriber:
     class: 'PrestaShopBundle\EventListener\ActionDispatcherLegacyHooksSubscriber'
@@ -30,10 +31,10 @@ services:
       - '@logger'
     tags:
       - { name: kernel.event_subscriber }
-      
-  prestashop.bundle.event_listener.admin.improved_b2b_feature_flag:
-    class: PrestaShopBundle\EventListener\Admin\ImprovedB2bFeatureFlagListener
-    arguments:
-      - '@PrestaShop\PrestaShop\Core\FeatureFlag\B2b\ImprovedB2bTabsToggler'
+
+  PrestaShopBundle\EventListener\Admin\ImprovedB2bFeatureFlagListener:
     tags:
-      - { name: doctrine.event_listener, event: postUpdate }
+      -
+        name: doctrine.orm.entity_listener
+        event: postUpdate
+        entity: PrestaShopBundle\Entity\FeatureFlag
