+++ src/PrestaShopBundle/Resources/config/services/bundle/event_subscriber.yml
@@ -1,0 +1,6 @@
+services:
+  _defaults:
+    public: true
+    autowire: true
+
+  PrestaShopBundle\EventSubscriber\UpdateShopModeFieldListener: ~
