+++ src/PrestaShopBundle/Service/Form/ImprovedB2bTabsToggler.php
@@ -4,13 +4,12 @@
  * docs/licenses/LICENSE.txt file that was distributed with this source code.
  */
 
-namespace PrestaShop\PrestaShop\Core\FeatureFlag\B2b;
+namespace PrestaShopBundle\Service\Form;
 
-use Doctrine\DBAL\ArrayParameterType;
-use Doctrine\DBAL\Connection;
 use PrestaShop\PrestaShop\Core\Feature\ShopModeFeature;
 use PrestaShop\PrestaShop\Core\FeatureFlag\FeatureFlagSettings;
 use PrestaShop\PrestaShop\Core\FeatureFlag\FeatureFlagStateCheckerInterface;
+use PrestaShopBundle\Entity\Repository\TabRepository;
 use Symfony\Contracts\Service\ResetInterface;
 
 final class ImprovedB2bTabsToggler
@@ -22,10 +21,9 @@ final class ImprovedB2bTabsToggler
     ];
 
     public function __construct(
-        private readonly Connection $connection,
         private readonly FeatureFlagStateCheckerInterface $featureFlagChecker,
         private readonly ShopModeFeature $shopModeFeature,
-        private readonly string $dbPrefix,
+        private readonly TabRepository $tabRepository,
     ) {
     }
 
@@ -44,20 +42,8 @@ final class ImprovedB2bTabsToggler
 
     private function setTabsActive(bool $active): void
     {
-        $sql = sprintf(
-            'UPDATE %stab SET active = :active WHERE class_name IN (:tabs)',
-            $this->dbPrefix
-        );
-
-        $this->connection->executeStatement(
-            $sql,
-            [
-                'active' => $active ? 1 : 0,
-                'tabs' => self::TAB_CLASS_NAMES,
-            ],
-            [
-                'tabs' => ArrayParameterType::STRING,
-            ]
-        );
+        foreach (self::TAB_CLASS_NAMES as $tabClassName) {
+            $this->tabRepository->changeStatusByClassName($tabClassName, $active);
+        }
     }
 }
